Ajustes varios, adicion de tabla de audit

+++ database/migrations/2025_01_20_151139_create_lotes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('lotes', function (Blueprint $table) {
             $table->id();
             $table->string('codigo')->unique();
-            $table->decimal('numero');
+            $table->integer('numero');
             $table->decimal('cantidad');
             $table->date('fechaVencimiento')->nullable();
             $table->foreignId('producto')->references('id')->on('productos');
